CHG: move exceptions to file
